refactor: move encryption logic to service layer
- Create CredentialDataService with buildCredentialCreateData function
- Update CredentialRepository to accept pre-prepared data with encryptedKey
- Revert app-store file changes (encryption handled by service layer)
- Revert crypto.ts changes (use existing symmetricEncrypt)
- Revert platform-libraries changes
