+++ packages/app-store/exchange2016calendar/api/add.ts
@@ -1,7 +1,7 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 import { z } from "zod";
 
-import { createEncryptedKey, symmetricEncrypt } from "@calcom/lib/crypto";
+import { symmetricEncrypt } from "@calcom/lib/crypto";
 import logger from "@calcom/lib/logger";
 import { defaultHandler } from "@calcom/lib/server/defaultHandler";
 import { defaultResponder } from "@calcom/lib/server/defaultResponder";
@@ -48,12 +48,8 @@ async function postHandler(req: NextApiRequest, res: NextApiResponse) {
       ...data,
     });
     await dav?.listCalendars();
-    const encryptedKey = createEncryptedKey(data.key);
     await prisma.credential.create({
-      data: {
-        ...data,
-        ...(encryptedKey && { encryptedKey }),
-      },
+      data,
     });
   } catch (reason) {
     logger.error("Could not add this exchange account", reason);
