+++ packages/app-store/ics-feedcalendar/api/add.ts
@@ -1,6 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import { createEncryptedKey, symmetricEncrypt } from "@calcom/lib/crypto";
+import { symmetricEncrypt } from "@calcom/lib/crypto";
 import logger from "@calcom/lib/logger";
 import prisma from "@calcom/prisma";
 
@@ -44,12 +44,8 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
         throw new Error(`Listed cals and URLs mismatch: ${listedCals.length} vs. ${urls.length}`);
       }
 
-      const encryptedKey = createEncryptedKey(data.key);
       await prisma.credential.create({
-        data: {
-          ...data,
-          ...(encryptedKey && { encryptedKey }),
-        },
+        data,
       });
     } catch (e) {
       logger.error("Could not add ICS feeds", e);
